+++ src/librustc_lint/types.rs
@@ -14,8 +14,8 @@ use middle::{infer};
 use middle::def_id::DefId;
 use middle::subst::Substs;
 use middle::ty::{self, Ty};
-use middle::const_eval::{eval_const_expr_partial, ConstVal};
-use middle::const_eval::EvalHint::ExprTypeChecked;
+use rustc_const_eval::eval::{eval_const_expr_partial, ConstVal};
+use rustc_const_eval::eval::EvalHint::ExprTypeChecked;
 use util::nodemap::{FnvHashSet};
 use lint::{LateContext, LintContext, LintArray};
 use lint::{LintPass, LateLintPass};
