+++ src/librustc_typeck/check/mod.rs
@@ -131,6 +131,8 @@ use rustc_front::hir::Visibility;
 use rustc_front::print::pprust;
 use rustc_back::slice;
 
+use rustc_const_eval::eval::eval_repeat_count;
+
 mod assoc;
 pub mod dropck;
 pub mod _match;
@@ -3548,7 +3550,7 @@ fn check_expr_with_unifier<'a, 'tcx, F>(fcx: &FnCtxt<'a, 'tcx>,
       }
       hir::ExprRepeat(ref element, ref count_expr) => {
         check_expr_has_type(fcx, &**count_expr, tcx.types.usize);
-        let count = fcx.tcx().eval_repeat_count(&**count_expr);
+        let count = eval_repeat_count(fcx.tcx(), &**count_expr);
 
         let uty = match expected {
             ExpectHasType(uty) => {
